Extract workspace composer APIs
